pyproject.toml
[project]
name = "flight_deviance_detection"
version = "0.1.0"
description = ""
authors = [{ name = "ruihan11"}]
requires-python = "==3.10.12"
dependencies = [
  "cachetools==6.1.0",
  "catboost==1.2.8",
  "click==8.2.1",
  "cloudpickle==3.1.1",
  "contourpy==1.3.2",
  "cuda-bindings==12.9.0",
  "cuda-python==12.9.0",
  "cudf-cu12==25.6.0",
  "cuml-cu12==25.6.0",
  "cupy-cuda12x==13.5.1",
  "cuvs-cu12==25.6.1",
  "cycler==0.12.1",
  "dask==2025.5.0",
  "dask-cuda==25.6.0",
  "dask-cudf-cu12==25.6.0",
  "distributed==2025.5.0",
  "distributed-ucxx-cu12==0.44.0",
  "fastrlock==0.8.3",
  "fonttools==4.59.0",
  "fsspec==2025.7.0",
  "graphviz==0.21",
  "importlib-metadata==8.7.0",
  "jinja2==3.1.6",
  "joblib==1.5.1",
  "kiwisolver==1.4.8",
  "libcudf-cu12==25.6.0",
  "libcuml-cu12==25.6.0",
  "libcuvs-cu12==25.6.1",
  "libkvikio-cu12==25.6.0",
  "libraft-cu12==25.6.0",
  "librmm-cu12==25.6.0",
  "libucx-cu12==1.18.1",
  "libucxx-cu12==0.44.0",
  "lightgbm==4.6.0",
  "llvmlite==0.44.0",
  "locket==1.0.0",
  "markdown-it-py==3.0.0",
  "markupsafe==3.0.2",
  "matplotlib==3.10.3",
  "mdurl==0.1.2",
  "msgpack==1.1.1",
  "narwhals==1.48.1",
  "numba==0.61.2",
  "numba-cuda==0.11.0",
  "numpy==2.2.6",
  "nvidia-cublas-cu12==12.9.1.4",
  "nvidia-cuda-nvcc-cu12==12.9.86",
  "nvidia-cuda-nvrtc-cu12==12.9.86",
  "nvidia-cufft-cu12==11.4.1.4",
  "nvidia-curand-cu12==10.3.10.19",
  "nvidia-cusolver-cu12==11.7.5.82",
  "nvidia-cusparse-cu12==12.5.10.65",
  "nvidia-ml-py==12.575.51",
  "nvidia-nccl-cu12==2.27.6",
  "nvidia-nvjitlink-cu12==12.9.86",
  "nvtx==0.2.12",
  "packaging==25.0",
  "pandas==2.2.3",
  "partd==1.4.2",
  "pillow==11.3.0",
  "plotly==6.2.0",
  "psutil==7.0.0",
  "pyarrow==19.0.1",
  "pygments==2.19.2",
  "pylibcudf-cu12==25.6.0",
  "pylibraft-cu12==25.6.0",
  "pynvjitlink-cu12==0.7.0",
  "pynvml==12.0.0",
  "pyparsing==3.2.3",
  "python-dateutil==2.9.0.post0",
  "pytz==2025.2",
  "pyyaml==6.0.2",
  "raft-dask-cu12==25.6.0",
  "rapids-dask-dependency==25.6.0",
  "rapids-logger==0.1.1",
  "rich==14.1.0",
  "rmm-cu12==25.6.0",
  "scikit-learn==1.7.1",
  "scipy==1.15.3",
  "six==1.17.0",
  "sortedcontainers==2.4.0",
  "tblib==3.1.0",
  "threadpoolctl==3.6.0",
  "toolz==1.0.0",
  "tornado==6.5.1",
  "tqdm==4.67.1",
  "treelite==4.4.1",
  "typing-extensions==4.14.1",
  "tzdata==2025.2",
  "ucx-py-cu12==0.44.0",
  "ucxx-cu12==0.44.0",
  "urllib3==2.5.0",
  "xgboost==3.0.2",
  "zict==3.0.0",
  "zipp==3.23.0"
]
